Move gradient() method from renderingUtil.ts to colorConverter.ts
