+++ visualization/app/codeCharta/ui/codeMap/rendering/renderingUtil.ts
@@ -1,6 +1,4 @@
 import {Node} from "../../../codeCharta.model";
-import {ColorConverter} from "../../../util/colorConverter";
-import {Color} from "three";
 
 export class RenderingUtil {
 
@@ -13,18 +11,5 @@ export class RenderingUtil {
         return max;
     }
 
-    public static gradient(startColor: string, endColor: string, steps: number): string[] {
-        let start: Color = ColorConverter.convertHexToColorObject(startColor)
-        let end: Color = ColorConverter.convertHexToColorObject(endColor)
-        let diff: Color = end.sub(start)
-        let stepsArray = [];
 
-        for(let i = 0; i <= steps; i++) {
-            let stepDiff = diff.clone().multiplyScalar(1 / steps * i)
-            let step = start.clone().add(stepDiff)
-            stepsArray[i] = ColorConverter.convertColorToHex(step)
-        }
-        return stepsArray
-
-    }
 }
